refactor(handlers): restructure handlers and service dependencies
- Refactored project to separate handler functions from service logic:
  - Created `URLShortenerHandler` and `URLStatHandler` in `internal/handler` for handling HTTP routes.
  - Moved service functions to `repository/url_shortener_repository.go` and injected dependencies for separation of concerns.
- Updated `redis.go` in `cache` to translate comments to English for consistency.
- Refactored `urlshortener.go`:
  - Replaced inline route functions with `URLShortenerHandler` and `URLStatHandler` instances.
  - Added routes for shortening, redirection, toggling URL state, and retrieving URL statistics.
- Adjusted `repository_test.go` to use `URLServiceImpl` from the `repository` package instead of `service`.
- Minor improvements in `url_stat_service.go` and `url_shortener_service.go`:
  - Standardized function naming, improved caching logic, and clarified comments.
