+++ internal/cache/redis.go
@@ -3,10 +3,9 @@ package cache
 import (
 	"context"
 	"errors"
-	"time"
-
 	"github.com/go-redis/redis/v8"
 	"github.com/reactivex/rxgo/v2"
+	"time"
 )
 
 var (
@@ -14,22 +13,22 @@ var (
 	ctx = context.Background()
 )
 
-// InitRedis inicializa el cliente de Redis
+// InitRedis initializes the Redis client
 func InitRedis(address, password string, db int) {
 	rdb = redis.NewClient(&redis.Options{
 		Addr:     address,
-		Password: password, // Deja vacío si no se utiliza autenticación
-		DB:       db,       // Número de la base de datos
+		Password: password, // Leave empty if no authentication is used
+		DB:       db,       // Database number
 	})
 
-	// Opcional: Prueba la conexión a Redis para asegurar que está funcionando
+	// Optional: Test the Redis connection to ensure it's working
 	_, err := rdb.Ping(ctx).Result()
 	if err != nil {
 		panic("Failed to connect to Redis: " + err.Error())
 	}
 }
 
-// SetURL almacena una URL en el caché de Redis de forma reactiva
+// SetURL stores a URL in the Redis cache reactively
 func SetURL(shortID string, originalURL string) rxgo.Observable {
 	return rxgo.Defer([]rxgo.Producer{func(_ context.Context, ch chan<- rxgo.Item) {
 		err := rdb.Set(ctx, shortID, originalURL, 24*time.Hour).Err()
@@ -41,12 +40,12 @@ func SetURL(shortID string, originalURL string) rxgo.Observable {
 	}})
 }
 
-// GetURL recupera una URL o valor del caché de Redis de forma reactiva
+// GetURL retrieves a URL or value from the Redis cache reactively
 func GetURL(key string) rxgo.Observable {
 	return rxgo.Defer([]rxgo.Producer{func(_ context.Context, ch chan<- rxgo.Item) {
 		result, err := rdb.Get(ctx, key).Result()
 		if errors.Is(err, redis.Nil) {
-			ch <- rxgo.Of("") // No encontrado en caché
+			ch <- rxgo.Of("") // Not found in cache
 		} else if err != nil {
 			ch <- rxgo.Error(err)
 		} else {
@@ -55,7 +54,7 @@ func GetURL(key string) rxgo.Observable {
 	}})
 }
 
-// DeleteURL elimina una URL del caché de Redis de forma reactiva
+// DeleteURL deletes a URL from the Redis cache reactively
 func DeleteURL(shortID string) rxgo.Observable {
 	return rxgo.Defer([]rxgo.Producer{func(_ context.Context, ch chan<- rxgo.Item) {
 		err := rdb.Del(ctx, shortID).Err()
@@ -67,22 +66,22 @@ func DeleteURL(shortID string) rxgo.Observable {
 	}})
 }
 
-// IncrementURLCounter incrementa el contador de accesos para una URL corta específica en Redis
+// IncrementURLCounter increments the access counter for a specific short URL in Redis
 func IncrementURLCounter(key string) error {
 	_, err := rdb.Incr(ctx, key).Result()
 	return err
 }
 
-// SetLastAccess establece la marca de tiempo del último acceso para una URL corta específica en Redis
+// SetLastAccess sets the timestamp of the last access for a specific short URL in Redis
 func SetLastAccess(key string, timestamp string) error {
 	return rdb.Set(ctx, key, timestamp, 0).Err()
 }
 
-// GetLastAccess recupera la marca de tiempo del último acceso para una URL corta específica de Redis
+// GetLastAccess retrieves the last access timestamp for a specific short URL from Redis
 func GetLastAccess(key string) (string, error) {
 	lastAccess, err := rdb.Get(ctx, key).Result()
 	if errors.Is(err, redis.Nil) {
-		// Si la clave no existe, devuelve una cadena vacía
+		// If the key does not exist, return an empty string
 		return "", nil
 	}
 	return lastAccess, err
