+++ internal/interfaces/url_shortener_service.go
@@ -4,7 +4,7 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-// URLShortenerServiceInterface define los métodos para el servicio de acortador de URLs
+// URLShortenerServiceInterface defines the methods for the URL shortener service
 type URLShortenerServiceInterface interface {
 	ShortenURLHandler(c *gin.Context)
 	RedirectURLHandler(c *gin.Context)
