+++ internal/repository/url_shortener_repository.go
@@ -1,4 +1,4 @@
-package service
+package repository
 
 import (
 	"context"
