+++ internal/service/url_shortener_service.go
@@ -1,97 +1,142 @@
-// internal/service/url_shortener_service.go
 package service
 
 import (
-	"context"
+	"crypto/md5"
+	"encoding/hex"
 	"errors"
+	"fmt"
 	"net/http"
-
-	"github.com/gin-gonic/gin"
-	"github.com/reactivex/rxgo/v2"
-
-	"urlshortener/internal/interfaces"
-	"urlshortener/internal/models"
-	"urlshortener/internal/request"
+	"urlshortener/internal/cache"
+	"urlshortener/internal/domain"
+	models2 "urlshortener/internal/models"
+	"urlshortener/internal/repository"
 )
 
-type URLShortenerService struct {
-	StatService interfaces.URLStatService
-}
+// URLServiceInstance URLService is an instance of the interface URLService which will be injected
+var URLServiceInstance repository.URLService
 
-// NewURLShortenerService crea una nueva instancia de URLShortenerService
-func NewURLShortenerService(statService interfaces.URLStatService) *URLShortenerService {
-	return &URLShortenerService{
-		StatService: statService,
-	}
-}
+// CreateShortURL generates a shortened URL and stores it in the database and cache
+func CreateShortURL(originalURL string) (string, error) {
 
-func (s *URLShortenerService) ShortenURLHandler(c *gin.Context) {
-	var req request.ShortenRequest
-	if err := c.ShouldBindJSON(&req); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid request payload"})
-		return
-	}
-
-	observable := rxgo.Just(req.OriginalURL)().
-		Map(func(_ context.Context, item interface{}) (interface{}, error) {
-			// Llama al servicio de acortamiento de URL
-			shortURL, err := CreateShortURL(item.(string))
-			return shortURL, err
-		})
-
-	result := <-observable.Observe()
-	if result.E != nil {
-		var apiErr *models.APIError
-		if errors.As(result.E, &apiErr) {
-			c.JSON(apiErr.Code, gin.H{"error": apiErr.Message})
-		} else {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to shorten URL"})
+	// Check if the original URL already exists in the database reactively
+	existsObservable := URLServiceInstance.FindURLByOriginal(originalURL)
+	existsResult := <-existsObservable.Observe()
+	if existsResult.E == nil && existsResult.V.(domain.URL).ShortURL != "" {
+		return existsResult.V.(domain.URL).ShortURL, &models2.APIError{
+			Code:    http.StatusConflict,
+			Message: "URL already exists",
 		}
-		return
 	}
 
-	c.JSON(http.StatusOK, gin.H{"short_url": result.V})
+	// Generate a unique ID (hash) for the shortened URL
+	shortID := generateShortID(originalURL)
+	shortURL := fmt.Sprintf("http://localhost:8080/%s", shortID)
+
+	// Create the URL structure
+	url := domain.URL{
+		ID:          shortID,
+		OriginalURL: originalURL,
+		ShortURL:    shortURL,
+		Enabled:     true,
+	}
+
+	// Save to MongoDB reactively
+	saveObservable := URLServiceInstance.SaveURL(url)
+	saveResult := <-saveObservable.Observe()
+	if saveResult.E != nil {
+		return "", saveResult.E
+	}
+
+	// Cache in Redis reactively
+	cacheObservable := cache.SetURL(shortID, originalURL)
+	cacheResult := <-cacheObservable.Observe()
+	if cacheResult.E != nil {
+		fmt.Printf("Error caching URL in Redis: %v\n", cacheResult.E) // Non-blocking error handling
+	}
+
+	return shortURL, nil
 }
 
-func (s *URLShortenerService) RedirectURLHandler(c *gin.Context) {
-	id := c.Param("id")
-
-	observable := rxgo.Just(id)().
-		Map(func(_ context.Context, item interface{}) (interface{}, error) {
-			// Llama al servicio para resolver la URL original
-			originalURL, err := ResolveURL(item.(string))
-			return originalURL, err
-		})
-
-	result := <-observable.Observe()
-	if result.E != nil || result.V == nil {
-		c.JSON(http.StatusNotFound, gin.H{"error": "URL not found"})
-		return
+// ResolveURL retrieves the original URL using the shortened ID
+func ResolveURL(shortID string) (string, error) {
+	// Try to get the URL from cache reactively
+	cacheObservable := cache.GetURL(shortID)
+	cacheResult := <-cacheObservable.Observe()
+	if cacheResult.E == nil && cacheResult.V.(string) != "" {
+		return cacheResult.V.(string), nil
 	}
 
-	// Registra el acceso en las estadísticas
-	recordObservable := s.StatService.RecordAccess(id)
-	recordResult := <-recordObservable.Observe()
-	if recordResult.E != nil {
-		// Puedes agregar logs aquí si lo deseas
+	// If not in cache, search in MongoDB reactively
+	dbObservable := URLServiceInstance.GetURL(shortID)
+	dbResult := <-dbObservable.Observe()
+	if dbResult.E != nil {
+		return "", errors.New("URL not found")
+	}
+	url := dbResult.V.(domain.URL)
+
+	// If disabled, return an error
+	if !url.Enabled {
+		return "", errors.New("URL is disabled")
 	}
 
-	// Redirige a la URL original
-	c.Redirect(http.StatusFound, result.V.(string))
+	// Cache for future requests reactively
+	cacheSaveObservable := cache.SetURL(shortID, url.OriginalURL)
+	cacheSaveResult := <-cacheSaveObservable.Observe()
+	if cacheSaveResult.E != nil {
+		fmt.Printf("Error caching URL in Redis: %v\n", cacheSaveResult.E) // Non-blocking error handling
+	}
+
+	return url.OriginalURL, nil
 }
 
-func (s *URLShortenerService) ToggleURLStateHandler(c *gin.Context) {
-	id := c.Param("id")
-	observable := rxgo.Just(id)().
-		Map(func(_ context.Context, item interface{}) (interface{}, error) {
-			// Llama al servicio para cambiar el estado de la URL
-			updated, err := ToggleURLState(item.(string))
-			return updated, err
-		})
-	result := <-observable.Observe()
-	if result.E != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to update URL state"})
-		return
+// ToggleURLState toggles the enabled/disabled state of the URL
+func ToggleURLState(shortID string) (bool, error) {
+	// Retrieve the URL from the database reactively
+	dbObservable := URLServiceInstance.GetURL(shortID)
+	dbResult := <-dbObservable.Observe()
+	if dbResult.E != nil {
+		return false, errors.New("URL not found")
 	}
-	c.JSON(http.StatusOK, gin.H{"success": result.V})
+	url := dbResult.V.(domain.URL)
+
+	// Toggle the enabled state
+	url.Enabled = !url.Enabled
+
+	// Save to MongoDB reactively
+	updateObservable := URLServiceInstance.UpdateURL(url)
+	updateResult := <-updateObservable.Observe()
+	if updateResult.E != nil {
+		return false, updateResult.E
+	}
+
+	// Remove or update in cache based on the new state
+	if url.Enabled {
+		cacheUpdateObservable := cache.SetURL(shortID, url.OriginalURL)
+		cacheUpdateResult := <-cacheUpdateObservable.Observe()
+		if cacheUpdateResult.E != nil {
+			fmt.Printf("Error updating cache in Redis: %v\n", cacheUpdateResult.E)
+			return false, cacheUpdateResult.E
+		}
+	} else {
+		cacheDeleteObservable := cache.DeleteURL(shortID)
+		cacheDeleteResult := <-cacheDeleteObservable.Observe()
+		if cacheDeleteResult.E != nil {
+			fmt.Printf("Error deleting from cache in Redis: %v\n", cacheDeleteResult.E)
+			return false, cacheDeleteResult.E
+		}
+	}
+	return url.Enabled, nil
+}
+
+// generateShortID generates a unique ID based on the original URL
+func generateShortID(originalURL string) string {
+	// Uses hashFunction to generate a hash and takes the first 6 characters
+	return hashFunction(originalURL)[:6]
+}
+
+// hashFunction generates an MD5 hash of the original URL and converts it to a string
+func hashFunction(originalURL string) string {
+	hashMd5 := md5.New()
+	hashMd5.Write([]byte(originalURL))
+	return hex.EncodeToString(hashMd5.Sum(nil))
 }
